Minor bugfix on uc_model.dart
This commits fixes a bug that was happening when trying to create a UC in the flutter web interface. The issue was that the backend (via gRPC) was returning the ects field as a string instead of a number, and the fromJson method was trying to cast it directly as a num, which failed. Now it handles both cases.